Move `Attribute` class into `protobuf` package
